sort: Special handling for citation-number in sorting
