+++ crates/proc/src/sort/output_format.rs
@@ -4,26 +4,25 @@
 //
 // Copyright © 2018 Corporation for Digital Scholarship
 
-use crate::String;
-use super::micro_html::micro_html_to_string;
-use super::{FormatCmd, LocalizedQuotes, OutputFormat};
-use crate::IngestOptions;
+use crate::prelude::*;
+use citeproc_io::output::{
+    micro_html::micro_html_to_string, FormatCmd, LocalizedQuotes, OutputFormat,
+};
+use citeproc_io::{IngestOptions, SmartString, lazy_char_transform_owned};
 
 use csl::{DisplayMode, Formatting};
 
 #[derive(Debug, Clone, PartialEq)]
-pub struct PlainText;
+pub struct SortStringFormat;
 
-impl Default for PlainText {
+impl Default for SortStringFormat {
     fn default() -> Self {
-        PlainText {}
+        SortStringFormat
     }
 }
 
-use crate::lazy_char_transform_owned;
-
 // We don't want these characters in a sort string
-fn remove_quotes(s: String) -> String {
+fn remove_quotes(s: SmartString) -> SmartString {
     lazy_char_transform_owned(s, |c: char| {
         if c == '\'' || c == '"' || c == ',' {
             None
@@ -34,10 +33,10 @@ fn remove_quotes(s: String) -> String {
     })
 }
 
-impl OutputFormat for PlainText {
-    type Input = String;
-    type Build = String;
-    type Output = String;
+impl OutputFormat for SortStringFormat {
+    type Input = SmartString;
+    type Build = SmartString;
+    type Output = SmartString;
     type BibMeta = ();
 
     fn meta(&self) -> Self::BibMeta {}
@@ -53,7 +52,7 @@ impl OutputFormat for PlainText {
     }
 
     #[inline]
-    fn text_node(&self, s: String, _: Option<Formatting>) -> Self::Build {
+    fn text_node(&self, s: SmartString, _: Option<Formatting>) -> Self::Build {
         s
     }
 
@@ -70,7 +69,7 @@ impl OutputFormat for PlainText {
                 a
             })
         } else {
-            String::new()
+            SmartString::new()
         }
     }
 
@@ -81,7 +80,7 @@ impl OutputFormat for PlainText {
         _f: Option<Formatting>,
     ) -> Self::Build {
         let std_string = nodes.join(delimiter);
-        String::from(std_string)
+        SmartString::from(std_string)
     }
 
     fn quoted(&self, b: Self::Build, _quotes: LocalizedQuotes) -> Self::Build {
@@ -121,9 +120,9 @@ impl OutputFormat for PlainText {
     }
 
     #[inline]
-    fn stack_preorder(&self, _s: &mut String, _stack: &[FormatCmd]) {}
+    fn stack_preorder(&self, _s: &mut SmartString, _stack: &[FormatCmd]) {}
     #[inline]
-    fn stack_postorder(&self, _s: &mut String, _stack: &[FormatCmd]) {}
+    fn stack_postorder(&self, _s: &mut SmartString, _stack: &[FormatCmd]) {}
     #[inline]
     fn tag_stack(&self, _formatting: Formatting, _: Option<DisplayMode>) -> Vec<FormatCmd> {
         Vec::new()
@@ -138,7 +137,7 @@ impl OutputFormat for PlainText {
     #[inline]
     fn apply_text_case(&self, build: &mut Self::Build, options: &IngestOptions) {
         let is_uppercase = !build.chars().any(|c| c.is_lowercase());
-        let string = std::mem::replace(build, String::new());
+        let string = std::mem::replace(build, SmartString::new());
         *build = options.transform_case(string, false, true, is_uppercase);
     }
 }
